#525 Extract BMS_CONFIG to own file to fix circular import
