+++ angular-client/src/utils/topic.utils.ts
@@ -1,4 +1,5 @@
-import { BMS_CONFIG, Chip, chipToString, Segment } from './bms.utils';
+import { BMS_CONFIG } from './bms.config';
+import { Chip, chipToString, Segment } from './bms.utils';
 
 export const alphaTemp = (segment: Segment, cell: number) => `BMS/PerCell/Alpha/${segment}/Therms/${cell}`;
 export const betaTemp = (segment: Segment, cell: number) => `BMS/PerCell/Beta/${segment}/Therms/${cell}`;
